@@ -1,5 +1,5 @@
 {
-	"name": "my-app",
+	"name": "virtualcoffeeio",
 	"version": "0.1.0",
 	"private": true,
 	"engines": {
@@ -9,8 +9,8 @@
 	"scripts": {
 		"preinstall": "npx only-allow pnpm",
 		"format": "npx prettier --write --ignore-unknown --list-different \"**/*\"",
-		"lint": "next lint",
-		"build-member-files": "node scripts/loadMemberFiles.js",
+		"lint": "eslint",
+		"build-member-files": "tsx scripts/loadMemberFiles.js",
 		"local-dev": "cross-env NODE_ENV=development pnpm netlify dev",
 		"build": "next build",
 		"start": "next start",
@@ -22,11 +22,13 @@
 		"build-member-files": "src/content/members/**/*.ts"
 	},
 	"dependencies": {
+		"@eslint/eslintrc": "^3.3.1",
+		"@eslint/js": "^9.37.0",
 		"@heroicons/react": "^2.2.0",
 		"@imgix/js-core": "^3.8.0",
 		"@mdx-js/loader": "^3.1.1",
 		"@mdx-js/react": "^3.1.1",
-		"@next/mdx": "15.5.4",
+		"@next/mdx": "15.5.5",
 		"@sindresorhus/slugify": "^2.2.1",
 		"@types/mdx": "^2.0.13",
 		"airtable": "^0.12.2",
@@ -34,7 +36,7 @@
 		"calendar-link": "^2.11.0",
 		"front-matter": "^4.0.2",
 		"graphql": "^16.11.0",
-		"graphql-request": "^6.1.0",
+		"graphql-request": "^7.2.0",
 		"hast-util-to-string": "^2.0.0",
 		"hastscript": "^8.0.0",
 		"leaflet": "^1.9.4",
@@ -42,12 +44,12 @@
 		"luxon": "^3.7.2",
 		"markdown-it": "^13.0.2",
 		"mdast-util-toc": "^7.1.0",
-		"next": "15.5.4",
+		"next": "15.5.5",
 		"npm-watch": "^0.13.0",
-		"react": "19.1.1",
-		"react-dom": "19.1.1",
-		"react-leaflet-cluster": "^3.1.1",
+		"react": "19.2.0",
+		"react-dom": "19.2.0",
 		"react-leaflet": "^5.0.0",
+		"react-leaflet-cluster": "^3.1.1",
 		"rehype-autolink-headings": "^6.1.1",
 		"rehype-highlight": "^6.0.0",
 		"rehype-sanitize": "^5.0.1",
@@ -62,26 +64,26 @@
 		"unified": "^10.1.2"
 	},
 	"devDependencies": {
-		"@faker-js/faker": "^8.4.1",
-		"@netlify/functions": "^2.8.2",
+		"@faker-js/faker": "^10.0.0",
 		"@netlify/plugin-nextjs": "^5.13.5",
-		"@types/leaflet": "^1.9.20",
+		"@types/leaflet": "^1.9.21",
 		"@types/leaflet.markercluster": "^1.5.6",
 		"@types/luxon": "^3.7.1",
-		"@types/node": "^24.7.0",
+		"@types/node": "^24.7.2",
 		"@types/react": "^19.2.2",
-		"@types/react-dom": "^19.2.1",
+		"@types/react-dom": "^19.2.2",
 		"@types/require-dir": "^1.0.4",
 		"@types/sanitize-html": "^2.16.0",
 		"all-contributors-cli": "^6.26.1",
 		"concurrently": "^9.2.1",
-		"cross-env": "^7.0.3",
+		"cross-env": "^10.1.0",
 		"esbuild": "^0.25.10",
-		"eslint": "^8.57.1",
-		"eslint-config-next": "15.5.4",
+		"eslint": "^9.37.0",
+		"eslint-config-next": "15.5.5",
 		"netlify-cli": "^23.9.1",
 		"postcss": "^8.5.6",
 		"prettier": "^3.6.2",
+		"sass": "^1.93.2",
 		"tailwindcss": "^3.4.18",
 		"tsx": "^4.20.6",
 		"typescript": "^5.9.3"
