@@ -1,20 +1,20 @@
-const { faker } = require('@faker-js/faker');
+import { faker } from '@faker-js/faker';
 
-module.exports = {
+const sponsors = {
 	organization: {
 		sponsorshipsAsMaintainer: {
 			nodes: [
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -27,14 +27,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -47,14 +47,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/matthew-${faker.internet.userName}`,
+						url: `https://github.com/matthew-${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: null,
@@ -67,14 +67,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -87,14 +87,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -107,14 +107,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -127,14 +127,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: null,
@@ -147,14 +147,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -167,14 +167,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -187,14 +187,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: null,
@@ -207,14 +207,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -227,14 +227,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -247,14 +247,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -267,14 +267,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/eclectic-${faker.internet.userName}`,
+						url: `https://github.com/eclectic-${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -287,14 +287,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -307,14 +307,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -327,14 +327,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -347,14 +347,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -367,14 +367,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -387,14 +387,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -407,14 +407,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -427,14 +427,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -447,14 +447,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: null,
@@ -467,14 +467,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -487,14 +487,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: faker.internet.url(),
@@ -507,14 +507,14 @@ module.exports = {
 				{
 					sponsorEntity: {
 						name: faker.person.fullName(),
-						login: faker.internet.userName(),
+						login: faker.internet.username(),
 						id: faker.string.uuid(),
 						avatarUrl_80: faker.image.avatar(),
 						avatarUrl_160: faker.image.avatar(),
 						avatarUrl_240: faker.image.avatar(),
 						avatarUrl_480: faker.image.avatar(),
 						avatarUrl_720: faker.image.avatar(),
-						url: `https://github.com/${faker.internet.userName}`,
+						url: `https://github.com/${faker.internet.username()}`,
 						description: faker.lorem.sentences(),
 						descriptionHTML: `<div>${faker.lorem.sentences()}</div>`,
 						websiteUrl: null,
@@ -614,3 +614,5 @@ module.exports = {
 		},
 	},
 };
+
+export default sponsors;
