@@ -1,4 +1,4 @@
-import { createMetaData } from '@/util/createMetaData.server';
+import Link from 'next/link';
 
 const handle = {
 	listTitle:
@@ -22,7 +22,7 @@ export default function Challenge() {
 		<>
 			<div className="alert alert-success">
 				This monthly challenge is complete. Congratulations! Please join us for
-				the <a href="/monthlychallenges/">next challenge</a>!
+				the <Link href="/monthlychallenges/">next challenge</Link>!
 			</div>
 			<h1>
 				<small>Monthly Challenge for March, 2021:</small> Get job ready!
