@@ -1,4 +1,4 @@
-/* eslint-disable react/no-unescaped-entities */
+ 
 import { Fragment } from 'react';
 import { createMetaData } from '@/util/createMetaData.server';
 import { getWritingChallengeData } from '@/data/monthlyChallenges/NaNoWriMo';
