@@ -1,3 +1,5 @@
+import Link from 'next/link';
+
 export function Breadcrumbs() {
 	return (
 		<div className="py-4">
@@ -5,7 +7,7 @@ export function Breadcrumbs() {
 				<nav aria-label="breadcrumb">
 					<ol className="breadcrumb">
 						<li className="breadcrumb-item">
-							<a href="/resources">Resources</a>
+							<Link href="/resources">Resources</Link>
 						</li>
 					</ol>
 				</nav>
