@@ -66,7 +66,7 @@ async function getMemberGithubData(
 		return mockMemberData(data);
 	}
 
-	let headers = {
+	const headers = {
 		Accept: 'application/vnd.github.v3+json',
 		Authorization: 'bearer ' + token,
 	};
@@ -103,10 +103,8 @@ async function getMemberGithubData(
 
 		const queries: string[] = [];
 		const githubData: GithubSearchUserLookup = {};
-
-		let i,
-			j,
-			chunk = 15;
+		const chunk = 15;
+		let i, j;
 		for (i = 0, j = data.length; i < j; i += chunk) {
 			queries.push(
 				`${data
@@ -148,8 +146,8 @@ function loadDirectory(
 	const list = Object.keys(memberList).map((key) => memberList[key]);
 
 	return list.sort(function (a, b) {
-		var nameA = a.github.toLowerCase(); // ignore upper and lowercase
-		var nameB = b.github.toLowerCase(); // ignore upper and lowercase
+		const nameA = a.github.toLowerCase(); // ignore upper and lowercase
+		const nameB = b.github.toLowerCase(); // ignore upper and lowercase
 		if (nameA < nameB) {
 			return -1;
 		}
