@@ -187,7 +187,7 @@ export async function getEpisode({
 	queryParams = '',
 }: {
 	slug: PodcastEpisode['slug'];
-	queryParams?: any;
+	queryParams?: string;
 }): Promise<PodcastEpisode | null> {
 	if (!(process.env.CMS_URL && process.env.CMS_TOKEN)) {
 		const fakeData = await import('./mocks/podcast.server');
